requirements.txt
click==8.0.1
Flask==2.0.1
Flask-JWT==0.3.2
itsdangerous==2.0.1
Jinja2==3.0.1
MarkupSafe==2.0.1
psycopg2==2.9.1
PyJWT==1.4.2
python-dotenv==0.17.1
Werkzeug==2.0.1
